Added setUser in dbUtil

+++ app/src/main/java/com/example/wallstreettycoon/useraccount/User.java
@@ -6,7 +6,7 @@ public class User { //Data model
     String userLastName;
     String userUsername;
     String userPassword;
-    Float userBalance;
+    Double userBalance;
 
     public User(String userPassword, String userUsername, String userLastName, String userFirstName, Integer userID) {
         this.userPassword = userPassword;
@@ -14,7 +14,7 @@ public class User { //Data model
         this.userLastName = userLastName;
         this.userFirstName = userFirstName;
         this.userID = userID;
-        this.userBalance = 0.0F;
+        this.userBalance = 0.0;
     }
 
     //user constructor with no user id
@@ -23,7 +23,7 @@ public class User { //Data model
         this.userUsername = userUsername;
         this.userLastName = userLastName;
         this.userFirstName = userFirstName;
-        this.userBalance = 0.0F;
+        this.userBalance = 0.0;
     }
 
     public int getUserID() {
@@ -66,11 +66,11 @@ public class User { //Data model
         this.userPassword = userPassword;
     }
 
-    public Float getUserBalance() {
+    public Double getUserBalance() {
         return userBalance;
     }
 
-    public void setUserBalance(Float userBalance) {
+    public void setUserBalance(Double userBalance) {
         this.userBalance = userBalance;
     }
 }
